Don't re-create channels if they already exist. Fixes #15
This is still incorrect behavior, since a channel is never removed from the list
when parted. However, because there's a problem elsewhere, the most pragmatic
thing to do here is keep the channels around and hope the irc client attached to
the ircd will ignore messages for channels it's no longer in.